changed: put application skin handling in separate class
