+++ xbmc/application/CMakeLists.txt
@@ -1,7 +1,9 @@
 set(SOURCES ApplicationPowerHandling.cpp
+            ApplicationSkinHandling.cpp
             ApplicationVolumeHandling.cpp)
 
 set(HEADERS ApplicationPowerHandling.h
+            ApplicationSkinHandling.h
             ApplicationVolumeHandling.h)
 
 core_add_library(application)
